config auth routes, protection login

+++ src/routes/memories.ts
@@ -2,83 +2,97 @@ import { FastifyInstance } from "fastify";
 import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
-
 export async function memoriesRoutes(app: FastifyInstance) {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
 
-  app.get('/memories', async () => {
-
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
-         createdAt: 'asc',
-      }
-    })
-  
-    return memories.map(memory => {
+        createdAt: "asc",
+      },
+    });
+
+    return memories.map((memory) => {
       return {
         id: memory.id,
         coverUrl: memory.coverUrl,
-        excerpt: memory.content.substring(0 , 115).concat('...'),
-      }
-    })
-  })
-  
-  app.get('/memories/:id', async (request) => {
-    
-    const paramsSchema = z.object({
-      id: z.string().uuid()
-    })
+        excerpt: memory.content.substring(0, 115).concat("..."),
+      };
+    });
+  });
 
-    const { id } = paramsSchema.parse(request.params)
+  app.get("/memories/:id", async (request, reply) => {
+    const paramsSchema = z.object({
+      id: z.string().uuid(),
+    });
+
+    const { id } = paramsSchema.parse(request.params);
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
-      }
-    })
+      },
+    });
 
-    return memory
-  })
-  
-  app.post('/memories', async (request) => {
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
+    return memory;
+  });
+
+  app.post("/memories", async (request) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    })
+    });
 
-    const {content, coverUrl, isPublic } = bodySchema.parse(request.body)
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: '8fa6b494-d446-4f81-94db-d08bcaccd788'
-      }
-    })
+        userId: request.user.sub,
+      },
+    });
 
-    return memory
-  })
+    return memory;
+  });
 
-  app.put('/memories/:id', async (request) => {
- 
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
-      id: z.string().uuid()
-    })
+      id: z.string().uuid(),
+    });
 
-    const { id } = paramsSchema.parse(request.params)
+    const { id } = paramsSchema.parse(request.params);
 
-    
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    })    
-    
-    const {content, coverUrl, isPublic } = bodySchema.parse(request.body)
+    });
 
-    const memory = await prisma.memory.update({
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -86,24 +100,33 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-      }
-    })
+      },
+    });
 
-    return memory
-  })
+    return memory;
+  });
 
-  app.delete('/memories/:id', async (request) => {
- 
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
-      id: z.string().uuid()
-    })
+      id: z.string().uuid(),
+    });
 
-    const { id } = paramsSchema.parse(request.params)
+    const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
         id,
       },
-    })
-  })
+    });
+  });
 }
